requirements.txt
django-jsonfield
Django==2.2.3
django-autocomplete-light==3.4.1
django-filter==2.2.0
djangorestframework==3.10.1
Markdown==3.1.1
Pillow==6.1.0
pytz==2019.1
sqlparse==0.3.0
